+++ src/registry/routes/helpers/get-components-history.js
